Fix build system issues and enable multi-architecture builds
- Fixed serial function redefinition conflicts by moving implementations from kernel.c to drivers/serial.c
- Added architecture-specific serial implementations for x86_64, aarch64, RISC-V, and ARM
- Fixed Makefile to exclude kernel_graphics.c to avoid kernel_main conflicts
- Successfully tested builds for all supported architectures:
  * aarch64 (ARM64) - RPi 3/4/5 and generic
  * arm (ARM32) - RPi 3/4/5 and generic
  * x86_64 (Intel/AMD 64-bit) - generic
  * riscv64 (RISC-V 64-bit) - generic
- All 8 build configurations now compile successfully
- Generated versioned kernel images in build-output/ directory
- Resolved missing boot file references and cross-compiler dependencies
